MV-4 > added creating habit screen

+++ app/src/main/java/com/drowsynomad/mirrovision/presentation/core/common/models/CategoryUI.kt
@@ -13,7 +13,7 @@ import kotlin.random.Random
 
 @Serializable
 data class CategoryUI(
-    val id: Long = Random.nextLong(),
+    val id: Int = Random.nextInt(),
     val name: String = emptyString(),
     val backgroundColor: CategoryMainColor = CategoryMainColor.Blue,
     @DrawableRes val iconRes: Int = R.drawable.ic_add,
